requirements.txt
# Core
python-dotenv==1.0.1
loguru==0.7.2
requests==2.31.0

# Scraping & PDF
beautifulsoup4==4.12.3
pypdf==4.2.0
playwright==1.44.0

# Gemini
google-generativeai==0.7.2

# Dashboard
streamlit==1.35.0
streamlit-extras==0.4.2

# Storage
supabase==2.4.0
plotly==5.22.0
reportlab==4.2.0
